Add an API error type that is returned as JSON in the event of an error
The APIError type is meant to provide descriptive, useful errors to API
clients. Simply returning an error string is often not useful. An error
string is the minimum, but other information such as which parameter
caused the request to fail should also be included. Only the message
field has been implemented in this commit as the other fields are
non-trivial to implement. Additional fields can be added to APIError
type in the future without breaking compatibility.

And update the API docs to remove the wrapping "error" object type.